feat: issue with post
