+++ src/lib/helper.ts
@@ -51,7 +51,7 @@ export function getFileType(fileType: string): 'image' | 'video' | undefined {
   return
 }
 
-export async function isValidUrl(url: string): Promise<boolean> {
+function isValidUrl(url: string) {
   // Check if the URL format is valid
   const urlPattern = new RegExp(
     '^(https?:\\/\\/)?' + // protocol
@@ -64,4 +64,19 @@ export async function isValidUrl(url: string): Promise<boolean> {
   )
 
   return urlPattern.test(url)
+}
+
+export function judgeImageAndVideoValid({
+  image_urls,
+  video_url,
+}: {
+  image_urls: string | boolean // input type
+  video_url: string | boolean
+}) {
+  const isImageUrlsValid = image_urls
+    ?.toString()
+    ?.split(',')
+    ?.every((url) => isValidUrl(url))
+  const isVideoUrlValid = isValidUrl(video_url?.toString())
+  return { isImageUrlsValid, isVideoUrlValid }
 }
